+++ docs/changelog.md
@@ -7,6 +7,7 @@
   - Support python 3.13 and 3.14.
   - Add Zarr backend support for AnnData.
   - Improve the precision of p values in `tl.motif_enrichment`.
+  - Add `pl.coverage` for quick visualization of signal coverage.
 
 ### Breaking Changes: